+++ tests/test_journal.py
@@ -10,6 +10,10 @@ def test_from_csv(accounts_file, txns_file, bassertions_file, budget_file):
     assert len(j.postings) == 8
     assert len(j.bassertions) == 6
     assert len(j.targets) == 4
+    a = j.account('Assets:Checking')
+    assert "Tag 1" not in a.tags
+    a2 = j.account('Assets')
+    assert "Tag 1" in a2.tags
 
 
 def test_check_balances(accounts_file, txns_file, bassertions_file):
@@ -107,7 +111,21 @@ def test_to_polars(accounts_file, txns_file):
     assert len(df) == 8
     expected_cols = ['Txn', 'Date', 'Account',  'Account short name',
                      'Account 1', 'Account 2', 'Account 3', 'Amount', 'Comment',
-                     'Stmt date', 'Stmt description', 'Number']
+                     'Stmt date', 'Stmt description', 'Number', 'Tag 1']
     assert len(df.columns) == len(expected_cols)
     for x in expected_cols:
         assert x in df.columns
+
+
+def test_conflicting_tags(accounts_file, txns_file):
+    j = Journal.from_csv(accounts=accounts_file, postings=txns_file)
+    ps = []
+    for p in j.postings:
+        p.tags['Tag 1'] = 'Conflict'
+        p.tags['Tag 1_acc'] = 'Conflict again'
+        ps.append(p)
+    df = j.to_polars(ps)
+
+    assert df['Tag 1'].unique().to_list() == ['Conflict']
+    assert df['Tag 1_acc'].unique().to_list() == ['Conflict again']
+    assert "Tag 1_acc2" in df.columns